+++ src/CoffeeShop.Application/CoffeeShop.Application/UseCase/Pedido/GetTotalValorPedido/IGetTotalValorPedidoUseCase.cs
@@ -4,6 +4,6 @@ namespace CoffeeShop.Application.UseCase.Pedido.GetTotalValorPedido
 {
     public interface IGetTotalValorPedidoUseCase
     {
-        long CalculateTotalValorPedido(List<PeiPedidoIten> listaPedido);
+        decimal CalculateTotalValorPedido(List<PeiPedidoIten> listaPedido);
     }
 }
